startup/org-mode.el: Fix clock-in/out keys.
Previously, running `org-clock-in' on an task which was not in the STARTED state
would clock-in, then the advice would set the state to STARTED, at which point
`org-clock-in' would run again, clocking out of the first interval and clocking
into a second.

I changed things so that C-c C-x C-i merely sets the state to STARTED and lets
the hook deal with setting the clock.

In this system, you never run clock-in or clock-out directly.

Signed-off-by: Daniel Hackney <dan@haxney.org>