Convert to startup files to `debian-run-directories'.
Files are now named `xxfilename.el' and are loaded in order of the first two
digits. This allows control over the order in which files are loaded, allowing
`init.el' to contain only the essentials.

Also added proper file headers to all startup files. They are a bit redundant,
but they make the files look all-pro.

Signed-off-by: Daniel Hackney <dan@haxney.org>